refactor : 뷰 홀더 인터페이스 이름 변경
