+++ app/src/main/java/woowacourse/shopping/ui/cart/cartAdapter/CartAdapter.kt
@@ -5,22 +5,22 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.model.CartProductUIModel
 import woowacourse.shopping.model.PageUIModel
-import woowacourse.shopping.ui.cart.cartAdapter.viewHolder.CartItemViewHolder
+import woowacourse.shopping.ui.cart.cartAdapter.viewHolder.CartProductViewHolder
 import woowacourse.shopping.ui.cart.cartAdapter.viewHolder.CartViewHolder
 import woowacourse.shopping.ui.cart.cartAdapter.viewHolder.NavigationViewHolder
 
-class CartAdapter(private val cartListener: CartListener) : RecyclerView.Adapter<CartItemViewHolder>() {
+class CartAdapter(private val cartListener: CartListener) : RecyclerView.Adapter<CartViewHolder>() {
     private val cartItems = mutableListOf<CartItemType>()
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CartItemViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CartViewHolder {
         return when (viewType) {
-            CartItemType.TYPE_ITEM -> CartViewHolder.from(parent, cartListener)
+            CartItemType.TYPE_ITEM -> CartProductViewHolder.from(parent, cartListener)
             CartItemType.TYPE_FOOTER -> NavigationViewHolder.from(parent, cartListener)
             else -> throw IllegalArgumentException("Invalid viewType")
         }
     }
 
-    override fun onBindViewHolder(holder: CartItemViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: CartViewHolder, position: Int) {
         return holder.bind(cartItems[position])
     }
