+++ app/src/main/java/woowacourse/shopping/ui/shopping/productAdapter/viewHolder/ReadMoreViewHolder.kt
@@ -10,7 +10,7 @@ class ReadMoreViewHolder private constructor(
     binding: ItemProductReadMoreBinding,
     listener: ProductsListener
 ) :
-    ItemViewHolder(binding.root) {
+    ShoppingViewHolder(binding.root) {
 
     init {
         binding.listener = listener
