+++ app/src/main/java/woowacourse/shopping/ui/shopping/productAdapter/viewHolder/ShoppingViewHolder.kt
@@ -4,6 +4,6 @@ import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.ui.shopping.productAdapter.ProductsItemType
 
-sealed class ItemViewHolder(view: View) : RecyclerView.ViewHolder(view) {
+sealed class ShoppingViewHolder(view: View) : RecyclerView.ViewHolder(view) {
     abstract fun bind(productItemType: ProductsItemType)
 }
